Fixing build.
git-svn-id: http://svn.code.sf.net/p/genoviz/code/trunk@11350 80d5a9c2-7d8c-4c9f-9369-4e03bca4d2bd

+++ igb/src/com/affymetrix/igb/util/ExportDialog.java
@@ -328,8 +328,8 @@ public class ExportDialog implements ExportConstants {
 			imageInfo.setHeight(component.getHeight());
 		}
 
-		widthSpinner.setValue((double) imageInfo.getWidth());
-		heightSpinner.setValue((double) imageInfo.getHeight());
+		widthSpinner.setValue(imageInfo.getWidth());
+		heightSpinner.setValue(imageInfo.getHeight());
 	}
 
 	/**
